backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0

# CORS middleware
fastapi-cors==0.0.6

# AI and NLP dependencies
openai==1.3.0
google-generativeai==0.3.1
google-cloud-speech==2.21.0
pyPDF2==3.0.1
python-docx==1.0.1

# Utilities
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
aiofiles==23.2.1